+++ src/event/equip_panel_run_character_status_thread.c
@@ -161,7 +161,7 @@ void equip_panel_run_character_status_thread(void) {
         screen->draw_offsets[1].y = g_equip_gfx_draw_offset_y;
         SetDrawOffset(&screen->draw_offsets[1], &screen->draw_offsets[1].x);
         equip_panel_set_primitive_colors(screen, thread);
-        equip_gfx_apply_menu_palette_for_mode((s32)&screen->numeric_frame, (s32*)thread);
+        equip_gfx_apply_menu_palette_for_mode(&screen->numeric_frame, (s32*)thread);
         if (frame < 13) {
             battle_gfx_draw_or_append_gpu_primitive((s32*)&screen->portrait.areas[1]);
         }
